+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.0.1
 - Add credit
+- Fix friend unread counts
 
 # 1.0.0
 - Initial Release